Time: 5 ms (78.73%) | Memory: 45.8 MB (85.96%) - LeetSync

+++ 9-palindrome-number/palindrome-number.java
@@ -2,7 +2,7 @@ class Solution {
     public boolean isPalindrome(int x) {
         if(x<0) return false;
         int y=x;
-        long rev=0;
+        int rev=0;
         while(x>0){
             int d=x%10;
             if(rev>Integer.MAX_VALUE/10) return false;
